requirements.txt

# required use of python3.11
# sudo dnf install python3.11 python3.11-devel

faster-whisper
pydub
python-dotenv
ffmpeg-python